style: Combine invitation related validation files into one file
